Move cluster sharding tests to unit suite
- Relocate TestComputeClusterSharding from smoke tests
  (test_runner_defs.py) to unit tests (test_generate_runners.py)
- Import compute_cluster_sharding in test_generate_runners.py
- Remove stale import from test_runner_defs.py

The sharding tests exercise compute_cluster_sharding which lives in
generate_runners.py, so they belong alongside the other unit tests
for that module rather than in the smoke test suite.
